Moved test_file to test folder.
